General changes to structure leftover
