+++ ethanalysis/fitting/models.py
