+++ ethanalysis/utils/__init__.py
@@ -1,0 +1,2 @@
+from ethanalysis.utils.colors import *
+from ethanalysis.utils.main import *
